Fixing issues of tour test for awesome_tshirt
- Fixing issue in the workflow. When trying to run this tour test I got an issue because the new order was not found. 
By debugging it the problem was that as we weren't clicking on the "Order" button, it wasn't calling to the other controller /awesome_tshirt/validate_order where the order is really created. Now it clicks the button and in the next endpoint it finds the "Make another order" button to switch to the /web. Therefore now the order is created
- Fixing typo in the class o_list_number. Otherwise despite the order was created it couldn't find and element in the UI with that content, as it was not a class